backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.10.0
supabase==2.3.0
playwright==1.40.0
beautifulsoup4==4.12.3
pandas==2.2.0
requests==2.31.0
pydantic==2.5.3
